pyproject.toml
[tool.poetry]
name = "courses_platform"
version = "0.1.0"
description = "REST backend for e-learning platform."
authors = ["Aleksei Panfilov"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^2.2"
django-rest-registration = "^0.5.0"
psycopg2-binary = "^2.8"
djangorestframework = "^3.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
